replace collect base tokens button
