+++ src/components/Asset/AssetActions/Download.tsx
@@ -7,37 +7,15 @@ import ButtonBuy from '@shared/ButtonBuy'
 import { secondsToString } from '@utils/ddo'
 import AlgorithmDatasetsListForCompute from './Compute/AlgorithmDatasetsListForCompute'
 import styles from './Download.module.css'
-import {
-  FileMetadata,
-  LoggerInstance,
-  ZERO_ADDRESS,
-  FixedRateExchange
-} from '@oceanprotocol/lib'
+import { FileMetadata, LoggerInstance, ZERO_ADDRESS } from '@oceanprotocol/lib'
 import { order } from '@utils/order'
 import { AssetExtended } from 'src/@types/AssetExtended'
 import { buyDtFromPool } from '@utils/pool'
 import { downloadFile } from '@utils/provider'
-import { getCollectTokensFeedback, getOrderFeedback } from '@utils/feedback'
 import { getOrderPriceAndFees } from '@utils/accessDetailsAndPricing'
 import { OrderPriceAndFees } from 'src/@types/Price'
 import { toast } from 'react-toastify'
-import { gql, OperationResult } from 'urql'
-import { fetchData, getQueryContext } from '@utils/subgraph'
-import { getOceanConfig } from '@utils/ocean'
-import { FixedRateExchanges } from 'src/@types/subgraph/FixedRateExchanges'
 
-const FixedRateExchangesQuery = gql`
-  query FixedRateExchanges($user: String, $exchangeId: String) {
-    fixedRateExchanges(where: { owner: $user, exchangeId: $exchangeId }) {
-      id
-      owner {
-        id
-      }
-      exchangeId
-      baseTokenBalance
-    }
-  }
-`
 export default function Download({
   asset,
   file,
@@ -61,9 +39,6 @@ export default function Download({
   const [isLoading, setIsLoading] = useState(false)
   const [isOwned, setIsOwned] = useState(false)
   const [validOrderTx, setValidOrderTx] = useState('')
-  const [isCollectLoading, setIsCollectLoading] = useState(false)
-  const [baseTokenBalance, setBaseTokenBalance] = useState(0)
-  const [collectStatusText, setCollectStatusText] = useState('')
   const [orderPriceAndFees, setOrderPriceAndFees] =
     useState<OrderPriceAndFees>()
   useEffect(() => {
@@ -103,29 +78,6 @@ export default function Download({
     accountId,
     isOwned
   ])
-
-  useEffect(() => {
-    if (!accountId || asset.nft.owner !== accountId) return
-    const queryContext = getQueryContext(Number(asset.chainId))
-
-    async function getBaseTokenBalance() {
-      const variables = {
-        user: accountId.toLowerCase(),
-        exchangeId: asset?.accessDetails?.addressOrId
-      }
-      const result: OperationResult<FixedRateExchanges> = await fetchData(
-        FixedRateExchangesQuery,
-        variables,
-        queryContext
-      )
-      result?.data?.fixedRateExchanges[0]?.baseTokenBalance
-        ? setBaseTokenBalance(
-            parseInt(result?.data?.fixedRateExchanges[0]?.baseTokenBalance)
-          )
-        : setBaseTokenBalance(0)
-    }
-    getBaseTokenBalance()
-  }, [accountId, asset?.accessDetails?.addressOrId, asset.chainId, asset.nft])
 
   async function handleOrderOrDownload() {
     setIsLoading(true)
@@ -174,37 +126,6 @@ export default function Download({
     setIsLoading(false)
   }
 
-  async function handleCollectTokens() {
-    setIsCollectLoading(true)
-    const config = getOceanConfig(asset?.chainId)
-    const fixed = new FixedRateExchange(web3, config.fixedRateExchangeAddress)
-    try {
-      setCollectStatusText(
-        getCollectTokensFeedback(
-          asset.accessDetails.baseToken?.symbol,
-          baseTokenBalance.toString()
-        )
-      )
-
-      const tx = await fixed.collectBT(
-        accountId,
-        asset?.accessDetails?.addressOrId
-      )
-
-      if (!tx) {
-        setIsCollectLoading(false)
-        return
-      }
-      setBaseTokenBalance(0)
-      return tx
-    } catch (error) {
-      LoggerInstance.log(error.message)
-      setIsCollectLoading(false)
-    } finally {
-      setIsCollectLoading(false)
-    }
-  }
-
   const PurchaseButton = () => (
     <ButtonBuy
       action="download"
@@ -227,26 +148,6 @@ export default function Download({
     />
   )
 
-  const CollectTokensButton = () => (
-    <ButtonBuy
-      action="collect"
-      onClick={handleCollectTokens}
-      disabled={baseTokenBalance === 0 || !baseTokenBalance}
-      hasPreviousOrder={false}
-      hasDatatoken={false}
-      dtSymbol={asset?.accessDetails?.baseToken.symbol}
-      dtBalance={baseTokenBalance.toString()}
-      datasetLowPoolLiquidity={false}
-      assetType=""
-      stepText={collectStatusText}
-      assetTimeout=""
-      isConsumable={false}
-      consumableFeedback=""
-      isBalanceSufficient={false}
-      isLoading={isCollectLoading}
-    />
-  )
-
   return (
     <aside className={styles.consume}>
       <div className={styles.info}>
@@ -263,13 +164,6 @@ export default function Download({
           {!isInPurgatory && <PurchaseButton />}
         </div>
       </div>
-
-      {asset?.accessDetails?.datatoken?.name !== '' &&
-        asset?.nft.owner === accountId && (
-          <div className={styles.collect}>
-            <CollectTokensButton />
-          </div>
-        )}
 
       {asset?.metadata?.type === 'algorithm' && (
         <AlgorithmDatasetsListForCompute
