+++ src/components/Asset/AssetContent/index.module.css
@@ -39,13 +39,17 @@
   padding: calc(var(--spacer) / 4) var(--spacer);
   border-top: 1px solid var(--border-color);
   border-bottom: 1px solid var(--border-color);
+  display: flex;
+  flex-direction: column;
 }
 
 .ownerActions a,
 .ownerActions button {
+  cursor: pointer;
   color: var(--color-secondary);
   margin-left: calc(var(--spacer) / 4);
   margin-right: calc(var(--spacer) / 4);
+  padding: calc(var(--spacer) / 4) var(--spacer);
 }
 
 .separator {
